+++ src/main/java/com/github/pagehelper/PageException.java
@@ -1,0 +1,22 @@
+package com.github.pagehelper;
+
+/**
+ * 分页插件异常
+ */
+public class PageException extends RuntimeException {
+    public PageException() {
+        super();
+    }
+
+    public PageException(String message) {
+        super(message);
+    }
+
+    public PageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public PageException(Throwable cause) {
+        super(cause);
+    }
+}
