+++ src/main/java/com/github/pagehelper/helper/BasePage.java
@@ -1,0 +1,161 @@
+package com.github.pagehelper.helper;
+
+import com.github.pagehelper.ISelect;
+import com.github.pagehelper.Page;
+import com.github.pagehelper.util.MetaObjectUtil;
+import org.apache.ibatis.reflection.MetaObject;
+
+import java.lang.reflect.Method;
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * 基础分页信息
+ *
+ * @author liuzh
+ */
+public abstract class BasePage {
+    protected static final ThreadLocal<Page> LOCAL_PAGE = new ThreadLocal<Page>();
+
+    //request获取方法
+    protected static Boolean hasRequest;
+    protected static Class<?> requestClass;
+    protected static Method getParameterMap;
+    protected static Map<String, String> PARAMS = new HashMap<String, String>(5);
+
+    static {
+        try {
+            requestClass = Class.forName("javax.servlet.ServletRequest");
+            getParameterMap = requestClass.getMethod("getParameterMap", new Class[]{});
+            hasRequest = true;
+        } catch (Throwable e) {
+            hasRequest = false;
+        }
+        PARAMS.put("pageNum", "pageNum");
+        PARAMS.put("pageSize", "pageSize");
+        PARAMS.put("count", "countSql");
+        PARAMS.put("orderBy", "orderBy");
+        PARAMS.put("reasonable", "reasonable");
+        PARAMS.put("pageSizeZero", "pageSizeZero");
+    }
+
+    /**
+     * 获取 Page 参数
+     *
+     * @return
+     */
+    protected static <T> Page<T> getLocalPage() {
+        return LOCAL_PAGE.get();
+    }
+
+    /**
+     * 设置 Page 参数
+     * @param page
+     */
+    protected static void setLocalPage(Page page) {
+        LOCAL_PAGE.set(page);
+    }
+
+    /**
+     * 移除本地变量
+     */
+    public static void clearPage() {
+        LOCAL_PAGE.remove();
+    }
+
+
+    /**
+     * 对象中获取分页参数
+     *
+     * @param params
+     * @return
+     */
+    protected static <T> Page<T> getPageFromObject(Object params) {
+        int pageNum;
+        int pageSize;
+        MetaObject paramsObject = null;
+        if (params == null) {
+            throw new NullPointerException("无法获取分页查询参数!");
+        }
+        if (hasRequest && requestClass.isAssignableFrom(params.getClass())) {
+            try {
+                paramsObject = MetaObjectUtil.forObject(getParameterMap.invoke(params, new Object[]{}));
+            } catch (Exception e) {
+                //忽略
+            }
+        } else {
+            paramsObject = MetaObjectUtil.forObject(params);
+        }
+        if (paramsObject == null) {
+            throw new NullPointerException("分页查询参数处理失败!");
+        }
+        Object orderBy = getParamValue(paramsObject, "orderBy", false);
+        boolean hasOrderBy = false;
+        if (orderBy != null && orderBy.toString().length() > 0) {
+            hasOrderBy = true;
+        }
+        try {
+            Object _pageNum = getParamValue(paramsObject, "pageNum", hasOrderBy ? false : true);
+            Object _pageSize = getParamValue(paramsObject, "pageSize", hasOrderBy ? false : true);
+            if (_pageNum == null || _pageSize == null) {
+                Page page = new Page();
+                page.setOrderBy(orderBy.toString());
+                page.setOrderByOnly(true);
+                return page;
+            }
+            pageNum = Integer.parseInt(String.valueOf(_pageNum));
+            pageSize = Integer.parseInt(String.valueOf(_pageSize));
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("分页参数不是合法的数字类型!");
+        }
+        Page page = new Page(pageNum, pageSize);
+        //count查询
+        Object _count = getParamValue(paramsObject, "count", false);
+        if (_count != null) {
+            page.setCount(Boolean.valueOf(String.valueOf(_count)));
+        }
+        //排序
+        if (hasOrderBy) {
+            page.setOrderBy(orderBy.toString());
+        }
+        //分页合理化
+        Object reasonable = getParamValue(paramsObject, "reasonable", false);
+        if (reasonable != null) {
+            page.setReasonable(Boolean.valueOf(String.valueOf(reasonable)));
+        }
+        //查询全部
+        Object pageSizeZero = getParamValue(paramsObject, "pageSizeZero", false);
+        if (pageSizeZero != null) {
+            page.setPageSizeZero(Boolean.valueOf(String.valueOf(pageSizeZero)));
+        }
+        return page;
+    }
+
+    /**
+     * 从对象中取参数
+     *
+     * @param paramsObject
+     * @param paramName
+     * @param required
+     * @return
+     */
+    protected static Object getParamValue(MetaObject paramsObject, String paramName, boolean required) {
+        Object value = null;
+        if (paramsObject.hasGetter(PARAMS.get(paramName))) {
+            value = paramsObject.getValue(PARAMS.get(paramName));
+        }
+        if (value != null && value.getClass().isArray()) {
+            Object[] values = (Object[]) value;
+            if (values.length == 0) {
+                value = null;
+            } else {
+                value = values[0];
+            }
+        }
+        if (required && value == null) {
+            throw new RuntimeException("分页查询缺少必要的参数:" + PARAMS.get(paramName));
+        }
+        return value;
+    }
+
+}
