+++ src/main/java/com/github/pagehelper/helper/dialect/HsqldbDialect.java
@@ -1,34 +1,30 @@
-package com.github.pagehelper.dialect;
+package com.github.pagehelper.helper.dialect;
 
 import com.github.pagehelper.Page;
-import com.github.pagehelper.util.SqlUtil;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.helper.HelperDialect;
 import org.apache.ibatis.cache.CacheKey;
 import org.apache.ibatis.session.RowBounds;
 
 /**
  * @author liuzh
  */
-public class MySqlDialect extends AbstractDialect {
-    public MySqlDialect(SqlUtil sqlUtil) {
-        super(sqlUtil);
-    }
+public class HsqldbDialect extends HelperDialect {
 
     @Override
     public String getPageSql(String sql, Page page, RowBounds rowBounds, CacheKey pageKey) {
-        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 14);
+        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 20);
         sqlBuilder.append(sql);
-        if (page.getStartRow() == 0) {
+        if (page.getPageSize() > 0) {
             sqlBuilder.append(" LIMIT ");
             sqlBuilder.append(page.getPageSize());
-        } else {
-            sqlBuilder.append(" LIMIT ");
+            pageKey.update(page.getPageSize());
+        }
+        if (page.getStartRow() > 0) {
+            sqlBuilder.append(" OFFSET ");
             sqlBuilder.append(page.getStartRow());
-            sqlBuilder.append(",");
-            sqlBuilder.append(page.getPageSize());
             pageKey.update(page.getStartRow());
         }
-        pageKey.update(page.getPageSize());
         return sqlBuilder.toString();
     }
-
 }
