+++ src/main/java/com/github/pagehelper/helper/dialect/SqlServer2012Dialect.java
@@ -1,32 +1,32 @@
-package com.github.pagehelper.dialect;
+package com.github.pagehelper.helper.dialect;
 
 import com.github.pagehelper.Page;
-import com.github.pagehelper.util.SqlUtil;
+import com.github.pagehelper.PageHelper;
 import org.apache.ibatis.cache.CacheKey;
 import org.apache.ibatis.session.RowBounds;
 
 /**
  * @author liuzh
  */
-public class HsqldbDialect extends AbstractDialect {
-    public HsqldbDialect(SqlUtil sqlUtil) {
-        super(sqlUtil);
-    }
+public class SqlServer2012Dialect extends SqlServerDialect {
 
     @Override
     public String getPageSql(String sql, Page page, RowBounds rowBounds, CacheKey pageKey) {
-        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 20);
+        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 14);
         sqlBuilder.append(sql);
-        if (page.getPageSize() > 0) {
-            sqlBuilder.append(" LIMIT ");
-            sqlBuilder.append(page.getPageSize());
-            pageKey.update(page.getPageSize());
-        }
         if (page.getStartRow() > 0) {
             sqlBuilder.append(" OFFSET ");
             sqlBuilder.append(page.getStartRow());
+            sqlBuilder.append(" ROWS ");
             pageKey.update(page.getStartRow());
+        }
+        if (page.getPageSize() > 0) {
+            sqlBuilder.append(" FETCH NEXT ");
+            sqlBuilder.append(page.getPageSize());
+            sqlBuilder.append(" ROWS ONLY");
+            pageKey.update(page.getPageSize());
         }
         return sqlBuilder.toString();
     }
+
 }
