+++ src/test/resources/hsqldb/mybatis-config-reasonable.xml
@@ -40,8 +40,7 @@
     </typeAliases>
 
     <plugins>
-        <plugin interceptor="com.github.pagehelper.PageHelper">
-            <property name="dialect" value="hsqldb"/>
+        <plugin interceptor="com.github.pagehelper.PageInterceptor">
             <property name="reasonable" value="true"/>
         </plugin>
     </plugins>
