+++ src/test/resources/hsqldb/mybatis-config.xml
@@ -40,11 +40,11 @@
     </typeAliases>
 
     <plugins>
-        <plugin interceptor="com.github.pagehelper.PageHelper">
+        <plugin interceptor="com.github.pagehelper.PageInterceptor">
             <!-- dialect参数可以忽略不写 -->
             <!-- 如果需要用自己的实现方法，可以将value写成类的全限定名称 -->
             <!-- 如下所示，该类必须继承Parser接口 -->
-            <!--<property name="dialect" value="hsqldb"/>-->
+            <!--<property name="helper" value="hsqldb"/>-->
             <!-- 支持通过Mapper接口参数来传递分页参数 -->
             <property name="supportMethodsArguments" value="true"/>
             <!-- 运行时自动判断数据库类型，支持多数据源DataSource分页 -->
